Atualiza e cria a interface de endereco service
